python/flowforge-jtbd-construction/pyproject.toml
[project]
name = "flowforge-jtbd-construction-starter"
version = "0.0.1"
description = "flowforge JTBD domain library — Construction (Bid, build, inspect)"
keywords = ["flowforge", "jtbd", "scaffold-only", "starter"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "flowforge-jtbd>=0.1.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flowforge_jtbd_construction"]

[tool.uv]
# E-46 / F-5: domain library not yet reviewed/rebranded. Workspace registers
# this dir for build discovery, but uv treats it as a virtual member (no
# wheel is built or installed) until E-48a/E-48b flips this flag back to
# true after security + product review.
package = false
